Sample for first and firstorDefault mathods
one more fix for Fib Gen - changed to Long

+++ simple-filtering/src/main/java/com/mogikanensoftware/rx/util/FibonacciGenerator.java
@@ -6,20 +6,20 @@ import java.util.List;
 
 public class FibonacciGenerator {
 
-	public static List<Integer> generateSequence(int j) {
-		List<Integer> result = new ArrayList<>();
+	public static List<Long> generateSequence(int j) {
+		List<Long> result = new ArrayList<>();
 		if (j < 0) {
 			throw new RuntimeException(String.format("Invalid parameter n=%d", j));
 		} else if (j == 0) {
 			System.out.println("Empty sequence");
 		} else if (j == 1) {
-			result.add(0);
+			result.add(0L);
 		} else if (j == 2) {
-			result.addAll(Arrays.asList(0, 1));
+			result.addAll(Arrays.asList(0L, 1L));
 		} else {
-			result.addAll(Arrays.asList(0, 1));
+			result.addAll(Arrays.asList(0L, 1L));
 			for (int i = 2; i < j; i++) {
-				int val = result.get(i - 1) + result.get(i - 2);
+				long val = result.get(i - 1) + result.get(i - 2);
 				result.add(val);
 			}
 		}
